#64 refactor : id 자료형 Long 으로 이관

+++ app/src/main/java/com/aoztg/greengrim/data/local/ChatDao.kt
@@ -9,7 +9,7 @@ import androidx.room.Query
 interface ChatDao {
 
     @Query("DELETE FROM unread_chat_data WHERE chatId = :chatId")
-    fun deleteUnReadChatData(chatId: Int)
+    fun deleteUnReadChatData(chatId: Long)
 
     @Insert(onConflict = REPLACE)
     fun addUnReadChatData(unReadChatEntity: UnReadChatEntity)
