+++ app/src/main/java/com/aoztg/greengrim/data/model/request/VerificationsRequest.kt
@@ -1,6 +1,6 @@
 package com.aoztg.greengrim.data.model.request
 
 data class VerificationsRequest(
-    val certificationId: Int,
+    val certificationId: Long,
     val response: Boolean
 )
