+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/home/mapper/UiHotChallengeMapper.kt
@@ -4,7 +4,7 @@ import com.aoztg.greengrim.data.model.response.HotChallengeInfo
 import com.aoztg.greengrim.presentation.ui.home.model.UiHotChallenge
 
 
-fun HotChallengeInfo.toUiHotChallenge(onItemClicked: (Int) -> Unit): UiHotChallenge {
+fun HotChallengeInfo.toUiHotChallenge(onItemClicked: (Long) -> Unit): UiHotChallenge {
     return UiHotChallenge(
         id = this.challengeInfo.id,
         imgUrl = this.challengeInfo.imgUrl,
