+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/main/MainActivity.kt
@@ -76,7 +76,7 @@ class MainActivity : BaseActivity<ActivityMainBinding>(ActivityMainBinding::infl
         if (intent.hasExtra("target")) {
             when (intent.getStringExtra("target")) {
                 "GRIM_DETAIL" -> {
-                    intent.getIntExtra("grimId", -1).let {
+                    intent.getLongExtra("grimId", -1).let {
                         navController.toGrimDetail(it)
                     }
                 }
@@ -85,7 +85,7 @@ class MainActivity : BaseActivity<ActivityMainBinding>(ActivityMainBinding::infl
                 }
 
                 "CREATE_NFT" -> {
-                    val grimId = intent.getIntExtra("grimId", -1)
+                    val grimId = intent.getLongExtra("grimId", -1)
                     intent.getStringExtra("grimUrl")?.let {
                         navController.toCreateNft(grimId, it) }
                 }
